requirements.txt
djangorestframework
graphene-django
django-filter
pytest-django
pytest-html
